fix(windows): move FileMonitorErrors to shared module and fix Bool to Int
- Move FileMonitorErrors enum to FileMonitorShared so it can be used
  by FileMonitorWindows without circular dependency
- Fix Bool to Int type mismatch for ReadDirectoryChangesW bWatchSubtree
  parameter (Windows BOOL is Int32, not Swift Bool)
