+++ Sources/FileMonitor/FileMonitor.swift
@@ -13,14 +13,6 @@ import FileMonitorLinux
 import FileMonitorWindows
 #endif
 
-/// Errors that `FileMonitor` can throw
-public enum FileMonitorErrors: Error {
-    case unsupported_os
-    case not_implemented_yet
-    case not_a_directory(url: URL)
-    case can_not_open(url: URL)
-}
-
 /// FileMonitor: Watch for file changes in a directory with a unified API on Linux and macOS.
 public struct FileMonitor: WatcherDelegate {
     private let fileChangeStream = AsyncStream.makeStream(of: FileChange.self)
